+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### NEXT
 
+Worker: Worker: Fix retransmissions, set proper marker bit
+([PR #N](https://github.com/versatica/mediasoup/pull/N)).
+
 ### 3.19.0
 
 - Node: Improve worker binary location detection ([PR #1603](https://github.com/versatica/mediasoup/pull/1603)).