+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### NEXT
 
 Worker: Worker: Fix retransmissions, set proper marker bit
-([PR #N](https://github.com/versatica/mediasoup/pull/N)).
+([PR #1606](https://github.com/versatica/mediasoup/pull/1606)).
 
 ### 3.19.0