+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### NEXT
 
-Worker: Worker: Fix retransmissions, set proper marker bit
+Worker: Fix retransmissions, set proper marker bit
 ([PR #1606](https://github.com/versatica/mediasoup/pull/1606)).
 
 ### 3.19.0